Package name fix and service refactoring
Fixed package name, refactored Main controller and ServiceManager
implementation
