+++ service-core/boot-service/src/main/java/com/mogikanensoftware/xml/service/BootServiceApplication.java
@@ -1,0 +1,12 @@
+package com.mogikanensoftware.xml.service;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BootServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(BootServiceApplication.class, args);
+	}
+}
