+++ service-core/boot-service/src/main/java/com/mogikanensoftware/xml/service/data/dao/ItemRepository.java
@@ -1,0 +1,11 @@
+package com.mogikanensoftware.xml.service.data.dao;
+
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+import com.mogikanensoftware.xml.service.data.entity.Item;
+
+@Repository
+public interface ItemRepository extends CrudRepository<Item,Long>{
+
+}
