+++ service-core/boot-service/src/main/java/com/mogikanensoftware/xml/service/data/entity/Result.java
@@ -1,0 +1,60 @@
+package com.mogikanensoftware.xml.service.data.entity;
+
+import java.io.Serializable;
+import java.sql.Date;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+
+@Entity
+public class Result implements Serializable{
+
+	private static final long serialVersionUID = 1L;
+
+	@Id
+	@GeneratedValue
+	private Long id;
+	
+	@Column(nullable=false)
+	private Date dateTime;
+
+	@Column(nullable=false)
+	private String fileName;
+
+	public Result() {
+		super();
+	}
+
+	public Result(Date dateTime, String fileName) {
+		super();
+		this.dateTime = dateTime;
+		this.fileName = fileName;
+	}
+	
+	public Date getDateTime() {
+		return dateTime;
+	}
+
+	public void setDateTime(Date dateTime) {
+		this.dateTime = dateTime;
+	}
+
+	public Long getId() {
+		return id;
+	}
+
+	public void setId(Long id) {
+		this.id = id;
+	}
+
+	public String getFileName() {
+		return fileName;
+	}
+
+	public void setFileName(String fileName) {
+		this.fileName = fileName;
+	}
+	
+}
