+++ service-core/boot-service/src/main/java/com/mogikanensoftware/xml/service/data/sm/ServiceManagerException.java
@@ -1,0 +1,15 @@
+package com.mogikanensoftware.xml.service.data.sm;
+
+public class ServiceManagerException extends Exception {
+
+	private static final long serialVersionUID = 1L;
+
+	public ServiceManagerException(String arg0, Throwable arg1) {
+		super(arg0, arg1);
+	}
+
+	public ServiceManagerException(String arg0) {
+		super(arg0);
+	}
+
+}
