+++ service-core/boot-service/src/main/java/com/mogikanensoftware/xml/service/data/transform/CustomTransformator.java
@@ -1,0 +1,13 @@
+package com.mogikanensoftware.xml.service.data.transform;
+
+import java.util.List;
+import java.util.Set;
+
+import com.mogikanensoftware.xml.service.data.entity.Item;
+import com.mogikanensoftware.xml.service.data.entity.Result;
+import com.mogikanensoftware.xml.utils.core.bean.ValidationInfoBean;
+
+public interface CustomTransformator {
+
+	List<Item> transform (Set<ValidationInfoBean> viBeans, Result result);
+}
