--- service-core/boot-service/src/main/java/mogikanensoftware/xml/service/BootServiceApplication.java
@@ -1,12 +1,0 @@
-package mogikanensoftware.xml.service;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BootServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(BootServiceApplication.class, args);
-	}
-}
