--- service-core/boot-service/src/main/java/mogikanensoftware/xml/service/data/entity/Item.java
@@ -1,171 +1,0 @@
-package mogikanensoftware.xml.service.data.entity;
-
-import java.io.Serializable;
-import java.sql.Date;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.ManyToOne;
-
-@Entity
-public class Item implements Serializable {
-
-	private static final long serialVersionUID = 1L;
-
-	@Id
-	@GeneratedValue
-	private Long id;
-
-	@Column(nullable = false)
-	private Date dateTime;
-
-	@Column(nullable = false)
-	private String itemType;
-
-	@Column
-	private String targetType;
-
-	@Column
-	private String targetName;
-
-	@Column
-	private String message;
-
-	@ManyToOne(fetch = FetchType.EAGER)
-	private Result result;
-	
-	public Item() {
-		super();
-	}
-
-	public Item(Date dateTime, String itemType, String targetType, String targetName, String message) {
-		super();
-		this.dateTime = dateTime;
-		this.itemType = itemType;
-		this.targetType = targetType;
-		this.targetName = targetName;
-		this.message = message;
-	}
-
-	public Long getId() {
-		return id;
-	}
-
-	public void setId(Long id) {
-		this.id = id;
-	}
-
-	public Date getDateTime() {
-		return dateTime;
-	}
-
-	public void setDateTime(Date dateTime) {
-		this.dateTime = dateTime;
-	}
-
-	public String getItemType() {
-		return itemType;
-	}
-
-	public void setItemType(String itemType) {
-		this.itemType = itemType;
-	}
-
-	public String getTargetType() {
-		return targetType;
-	}
-
-	public void setTargetType(String targetType) {
-		this.targetType = targetType;
-	}
-
-	public String getTargetName() {
-		return targetName;
-	}
-
-	public void setTargetName(String targetName) {
-		this.targetName = targetName;
-	}
-
-	public String getMessage() {
-		return message;
-	}
-
-	public void setMessage(String message) {
-		this.message = message;
-	}
-
-	public Result getResult() {
-		return result;
-	}
-
-	public void setResult(Result result) {
-		this.result = result;
-	}
-
-	
-	@Override
-	public String toString() {
-		return "Item [id=" + id + ", dateTime=" + dateTime + ", itemType=" + itemType + ", targetType=" + targetType
-				+ ", targetName=" + targetName + ", message=" + message + "]";
-	}
-
-	@Override
-	public int hashCode() {
-		final int prime = 31;
-		int result = 1;
-		result = prime * result + ((dateTime == null) ? 0 : dateTime.hashCode());
-		result = prime * result + ((id == null) ? 0 : id.hashCode());
-		result = prime * result + ((itemType == null) ? 0 : itemType.hashCode());
-		result = prime * result + ((message == null) ? 0 : message.hashCode());
-		result = prime * result + ((targetName == null) ? 0 : targetName.hashCode());
-		result = prime * result + ((targetType == null) ? 0 : targetType.hashCode());
-		return result;
-	}
-
-	@Override
-	public boolean equals(Object obj) {
-		if (this == obj)
-			return true;
-		if (obj == null)
-			return false;
-		if (getClass() != obj.getClass())
-			return false;
-		Item other = (Item) obj;
-		if (dateTime == null) {
-			if (other.dateTime != null)
-				return false;
-		} else if (!dateTime.equals(other.dateTime))
-			return false;
-		if (id == null) {
-			if (other.id != null)
-				return false;
-		} else if (!id.equals(other.id))
-			return false;
-		if (itemType == null) {
-			if (other.itemType != null)
-				return false;
-		} else if (!itemType.equals(other.itemType))
-			return false;
-		if (message == null) {
-			if (other.message != null)
-				return false;
-		} else if (!message.equals(other.message))
-			return false;
-		if (targetName == null) {
-			if (other.targetName != null)
-				return false;
-		} else if (!targetName.equals(other.targetName))
-			return false;
-		if (targetType == null) {
-			if (other.targetType != null)
-				return false;
-		} else if (!targetType.equals(other.targetType))
-			return false;
-		return true;
-	}
-
-}
