--- service-core/boot-service/src/main/java/mogikanensoftware/xml/service/data/sm/ServiceManagerException.java
@@ -1,15 +1,0 @@
-package mogikanensoftware.xml.service.data.sm;
-
-public class ServiceManagerException extends Exception {
-
-	private static final long serialVersionUID = 1L;
-
-	public ServiceManagerException(String arg0, Throwable arg1) {
-		super(arg0, arg1);
-	}
-
-	public ServiceManagerException(String arg0) {
-		super(arg0);
-	}
-
-}
