+++ service-core/boot-service/src/test/java/com/mogikanensoftware/xml/service/BootServiceApplicationTests.java
@@ -1,0 +1,19 @@
+package com.mogikanensoftware.xml.service;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+import com.mogikanensoftware.xml.service.data.sm.impl.TestAppConfig;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest(classes = TestAppConfig.class)
+public class BootServiceApplicationTests {
+
+	@Test
+	public void contextLoads() {
+		
+	}
+
+}
