--- service-core/boot-service/src/test/java/mogikanensoftware/xml/service/BootServiceApplicationTests.java
@@ -1,17 +1,0 @@
-package mogikanensoftware.xml.service;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest(classes = TestAppConfig.class)
-public class BootServiceApplicationTests {
-
-	@Test
-	public void contextLoads() {
-		
-	}
-
-}
