+++ service-core/boot-service/src/test/resources/log4j2.xml
@@ -10,6 +10,10 @@
 			level="DEBUG" additivity="false">
 			<AppenderRef ref="Console" />
 		</Logger>
+		<Logger name="com.mogikanensoftware.xml.service.data.sm.impl.ServiceManagerImpl"
+			level="DEBUG" additivity="false">
+			<AppenderRef ref="Console" />
+		</Logger>
 		<Root level="INFO">
 			<AppenderRef ref="Console" />
 		</Root>
